+++ dev/devicelab/bin/tasks/build_aar_module_test.dart
@@ -152,7 +152,7 @@ Future<void> main() async {
 
       section('Check assets in release AAR');
 
-      checkItContains<String>(
+      checkCollectionContains<String>(
         <String>[
           ...flutterAssets,
           // AOT snapshots
@@ -278,7 +278,7 @@ Future<void> main() async {
         'flutter_debug-1.0.aar',
       ));
 
-      checkItContains<String>(<String>[
+      checkCollectionContains<String>(<String>[
         ...flutterAssets,
         ...debugAssets,
       ], debugAar);
